Add some missing dependencies in Centos 7
Some required items in both pip and yum are missing from the base
centos 7 install, and are not handled during the deployment or
build. In addition, the iptables module needs to be inserted
and the libvirtd service started. While most of the python deps are
python 34, some parts of the build process are running in python2.